+++ mrlang/accept_language.go
@@ -51,7 +51,7 @@ func ParseAcceptLanguage(s string) []string {
 			if len(lang) > 2 {
 				addLang(lang[0:2] + "_" + lang[3:]) // ru-RU -> ru_RU
 			} else {
-				if addLang(lang) {
+				if addLang(lang) { // ru + ru_RU
 					addLang(lang + "_" + strings.ToUpper(lang))
 				}
 			}